Expanded important files.
We can upload the whole chatterbox Android Studio file so everyone can
download it and open the app within Android Studio. But then if we make
changes to our .java or .xml files, and if we copy the whole Chatterbox
file into our Github directory, the github cannot look through each or
the directories in the file to see what has changed and just download
those. Thus, every time we change any of the .java of .xml files, we
should copy their codes INDIVIDUALLY to the files I laid out here.